Testing gradle scripts; refactoring packages in test to be more clear
