+++ test/unit/SuitableMethodFinderTest.java
@@ -1,4 +1,4 @@
-package test;
+package unit;
 
 import java.io.File;
 import java.io.IOException;
