+++ test/unit/transformer/visitors/TransformVisitorTest.java
@@ -1,4 +1,4 @@
-package test.transformer.visitors;
+package unit.transformer.visitors;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
